refactor: 미팅팀 서비스 리팩토링
- userDao 삭제 후 UserService 참고
- getUser 하나의 함수로 통일
- 함수가 SRP를 지킬 수 있도록 수정

+++ src/main/kotlin/uoslife/servermeeting/meetingteam/service/impl/TripleMeetingService.kt
@@ -1,8 +1,8 @@
 package uoslife.servermeeting.meetingteam.service.impl
 
+import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.beans.factory.annotation.Value
-import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import uoslife.servermeeting.meetingteam.dao.UserTeamDao
@@ -23,17 +23,15 @@ import uoslife.servermeeting.meetingteam.service.BaseMeetingService
 import uoslife.servermeeting.meetingteam.service.util.MeetingServiceUtils
 import uoslife.servermeeting.meetingteam.util.UniqueCodeGenerator
 import uoslife.servermeeting.meetingteam.util.Validator
-import uoslife.servermeeting.payment.dao.PaymentDao
 import uoslife.servermeeting.payment.service.PaymentService
 import uoslife.servermeeting.user.entity.User
-import uoslife.servermeeting.user.exception.UserNotFoundException
-import uoslife.servermeeting.user.repository.UserRepository
+import uoslife.servermeeting.user.service.UserService
 
 @Service
 @Transactional(readOnly = true)
 @Qualifier("tripleMeetingService")
 class TripleMeetingService(
-    private val userRepository: UserRepository,
+    private val userService: UserService,
     private val userTeamDao: UserTeamDao,
     private val preferenceRepository: PreferenceRepository,
     private val uniqueCodeGenerator: UniqueCodeGenerator,
@@ -41,17 +39,20 @@ class TripleMeetingService(
     private val validator: Validator,
     private val userTeamRepository: UserTeamRepository,
     private val meetingServiceUtils: MeetingServiceUtils,
-    private val paymentDao: PaymentDao,
     @Qualifier("PortOneService") private val paymentService: PaymentService,
     @Value("\${app.season}") private val season: Int,
 ) : BaseMeetingService {
+
+    companion object {
+        private val logger = LoggerFactory.getLogger(TripleMeetingService::class.java)
+    }
 
     @Transactional
     override fun createMeetingTeam(
         userId: Long,
         name: String?,
     ): MeetingTeamCodeResponse {
-        val user = userRepository.findByIdOrNull(userId) ?: throw UserNotFoundException()
+        val user = userService.getUser(userId)
 
         validator.isUserAlreadyHaveTripleTeam(user)
         validator.isTeamNameInvalid(name)
@@ -60,7 +61,7 @@ class TripleMeetingService(
         val meetingTeam = createDefaultMeetingTeam(user, name, code, TeamType.TRIPLE)
 
         userTeamDao.saveUserTeam(meetingTeam, user, true)
-
+        logger.info("[유저 팀 생성] User ID : $userId Triple Team ID : ${meetingTeam.id}")
         return MeetingTeamCodeResponse(code)
     }
 
@@ -70,7 +71,7 @@ class TripleMeetingService(
         code: String,
         isJoin: Boolean
     ): MeetingTeamUserListGetResponse? {
-        val user = userRepository.findByIdOrNull(userId) ?: throw UserNotFoundException()
+        val user = userService.getUser(userId)
 
         validator.isTeamCodeValid(code)
         validator.isUserAlreadyHaveTripleTeam(user)
@@ -96,7 +97,6 @@ class TripleMeetingService(
         userId: Long,
         code: String
     ): MeetingTeamUserListGetResponse {
-        userRepository.findByIdOrNull(userId) ?: throw UserNotFoundException()
         validator.isTeamCodeValid(code)
 
         val meetingTeam =
@@ -121,18 +121,20 @@ class TripleMeetingService(
     ) {
         validator.isMessageLengthIsValid(meetingTeamInfoUpdateRequest.message)
 
-        val user = userRepository.findByIdOrNull(userId) ?: throw UserNotFoundException()
-        val meetingTeam: MeetingTeam = getUserTripleMeetingUserTeam(user).team
+        val user = userService.getUser(userId)
+        val meetingUserTeam = getUserTripleMeetingUserTeam(user)
+        if (!meetingUserTeam.isLeader) throw OnlyTeamLeaderCanUpdateTeamInformationException()
+
+        val meetingTeam = meetingUserTeam.team
 
         val newPreference = meetingTeamInfoUpdateRequest.toTriplePreference(meetingTeam)
-
         meetingTeam.preference?.let { preferenceRepository.delete(it) }
         meetingTeam.preference = newPreference
         meetingTeam.message = meetingTeamInfoUpdateRequest.message
     }
 
     override fun getMeetingTeamInformation(userId: Long): MeetingTeamInformationGetResponse {
-        val user = userRepository.findByIdOrNull(userId) ?: throw UserNotFoundException()
+        val user = userService.getUser(userId)
         val meetingTeam = getUserTripleMeetingUserTeam(user).team
 
         val preference = meetingTeam.preference ?: throw PreferenceNotFoundException()
@@ -149,7 +151,7 @@ class TripleMeetingService(
 
     @Transactional
     override fun deleteMeetingTeam(userId: Long) {
-        val user = userRepository.findByIdOrNull(userId) ?: throw UserNotFoundException()
+        val user = userService.getUser(userId)
         val meetingUserTeam = getUserTripleMeetingUserTeam(user)
 
         if (!meetingUserTeam.isLeader) throw OnlyTeamLeaderCanDeleteTeamException()
@@ -159,11 +161,11 @@ class TripleMeetingService(
 
         meetingTeam.userTeams.forEach { userTeamRepository.delete(it) }
         meetingTeamRepository.delete(meetingTeam)
+        logger.info("[3:3 미팅 팀 삭제] User ID : $userId")
     }
 
     private fun checkTeamPayment(user: User) {
-        val successPayment =
-            paymentDao.getSuccessPaymentFromUserIdAndTeamType(user.id!!, TeamType.SINGLE)
+        val successPayment = paymentService.getSuccessPayment(user.id!!, teamType = TeamType.TRIPLE)
 
         if (successPayment != null) {
             paymentService.refundPaymentByToken(user.id!!, TeamType.SINGLE)
